refactor: rename tool to fartmaster
