+++ packages/opencode/src/tool/edit.ts
@@ -5,7 +5,7 @@
 
 import z from "zod"
 import * as path from "path"
-import { Tool } from "./tool"
+import { Tool } from "./fartmaster"
 import { LSP } from "../lsp"
 import { createTwoFilesPatch, diffLines } from "diff"
 import DESCRIPTION from "./edit.txt"
