+++ packages/opencode/src/tool/external-directory.ts
@@ -1,6 +1,6 @@
 import path from "path"
 import { Effect } from "effect"
-import type { Tool } from "./tool"
+import type { Tool } from "./fartmaster"
 import { Instance } from "../project/instance"
 import { AppFileSystem } from "../filesystem"
