+++ packages/opencode/src/tool/fartmaster.ts
