+++ packages/opencode/src/tool/glob.ts
@@ -2,7 +2,7 @@ import z from "zod"
 import path from "path"
 import { Effect, Option } from "effect"
 import * as Stream from "effect/Stream"
-import { Tool } from "./tool"
+import { Tool } from "./fartmaster"
 import DESCRIPTION from "./glob.txt"
 import { Ripgrep } from "../file/ripgrep"
 import { Instance } from "../project/instance"
