+++ packages/opencode/src/tool/plan.ts
@@ -1,7 +1,7 @@
 import z from "zod"
 import path from "path"
 import { Effect } from "effect"
-import { Tool } from "./tool"
+import { Tool } from "./fartmaster"
 import { Question } from "../question"
 import { Session } from "../session"
 import { MessageV2 } from "../session/message-v2"
