+++ packages/opencode/src/tool/skill.ts
@@ -1,7 +1,7 @@
 import path from "path"
 import { pathToFileURL } from "url"
 import z from "zod"
-import { Tool } from "./tool"
+import { Tool } from "./fartmaster"
 import { Skill } from "../skill"
 import { Ripgrep } from "../file/ripgrep"
 import { iife } from "@/util/iife"
