+++ packages/opencode/src/tool/websearch.ts
@@ -1,7 +1,7 @@
 import z from "zod"
 import { Effect } from "effect"
 import { HttpClient } from "effect/unstable/http"
-import { Tool } from "./tool"
+import { Tool } from "./fartmaster"
 import * as McpExa from "./mcp-exa"
 import DESCRIPTION from "./websearch.txt"
